Made a project file decomposition.
