+++ main/inc/api_constants.h
@@ -1,0 +1,51 @@
+/* esp-idf-telegram-bot
+ *
+ * Author: antusystem
+ * e-mail: aleantunes95@gmail.com
+ * Date: 11-01-2020
+ * MIT License
+ * As it is described in the readme file
+ *
+*/
+
+/*HTTP buffer*/
+#define MAX_HTTP_RECV_BUFFER 1024
+#define MAX_HTTP_OUTPUT_BUFFER 2048
+
+/* TAGs for the system*/
+static const char *TAG = "HTTP_CLIENT Handler";
+static const char *TAG1 = "wifi station";
+static const char *TAG2 = "Sending getMe";
+static const char *TAG3 = "Sending sendMessage";
+
+/*WIFI configuration*/
+#define ESP_WIFI_SSID      "YOUR_SSID_NAME"
+#define ESP_WIFI_PASS      "YOUR_SSID_PASSWORD"
+#define ESP_MAXIMUM_RETRY  10
+
+/*Telegram configuration*/
+#define TOKEN "YOUR_TOKEN"
+char url_string[512] = "https://api.telegram.org/bot";
+// Using in the task strcat(url_string,TOKEN)); the main direct from the url will be in url_string
+//The chat id that will receive the message
+#define chat_ID1 "@GROUP_NAME"
+#define chat_ID2 "123456789"
+
+/* The event group allows multiple bits for each event, but we only care about two events:
+ * - we are connected to the AP with an IP
+ * - we failed to connect after the maximum amount of retries */
+#define WIFI_CONNECTED_BIT BIT0
+#define WIFI_FAIL_BIT      BIT1
+
+int s_retry_num = 0;
+
+
+/* Root cert for extracted from:
+ *
+ * https://github.com/witnessmenow/Universal-Arduino-Telegram-Bot/blob/master/src/TelegramCertificate.h
+
+   To embed it in the app binary, the PEM file is named
+   in the component.mk COMPONENT_EMBED_TXTFILES variable.
+*/
+extern const char telegram_certificate_pem_start[] asm("_binary_telegram_certificate_pem_start");
+extern const char telegram_certificate_pem_end[]   asm("_binary_telegram_certificate_pem_end");
