+++ main/inc/network_handler.h
@@ -1,0 +1,12 @@
+/* esp-idf-telegram-bot
+ *
+ * Author: antusystem
+ * e-mail: aleantunes95@gmail.com
+ * Date: 11-01-2020
+ * MIT License
+ * As it is described in the readme file
+ *
+*/
+
+void wifi_init_sta(void);
+void http_test_task(void *pvParameters);
